+++ common/src/main/java/whocraft/tardis_refined/common/mixin/PlayerEntityMixin.java
@@ -7,6 +7,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import whocraft.tardis_refined.common.GravityClient;
 import whocraft.tardis_refined.common.GravityUtil;
 
 @Mixin(Player.class)
@@ -15,7 +16,8 @@ public class PlayerEntityMixin {
     @Inject(method = "travel(Lnet/minecraft/world/phys/Vec3;)V", at = @At("TAIL"), cancellable = true)
     private void move(Vec3 vec3, CallbackInfo info) {
         Player player = (Player) (Object) this;
-        GravityUtil.moveGravity(player, info);
+        if(!player.level().isClientSide) return;
+        GravityClient.moveGravity(player, info);
     }
