+++ forge/src/main/java/whocraft/tardis_refined/neoforge/ClientForgeBus.java
@@ -9,6 +9,7 @@ import net.neoforged.neoforge.event.TickEvent;
 import whocraft.tardis_refined.TardisRefined;
 import whocraft.tardis_refined.client.GravityOverlay;
 import whocraft.tardis_refined.client.TardisClientData;
+import whocraft.tardis_refined.client.TardisClientLogic;
 
 @Mod.EventBusSubscriber(modid = TardisRefined.MODID, bus = Mod.EventBusSubscriber.Bus.FORGE, value = Dist.CLIENT)
 public class ClientForgeBus {
@@ -19,7 +20,7 @@ public class ClientForgeBus {
             return;
         }
 
-        TardisClientData.tickClientData(Minecraft.getInstance());
+        TardisClientLogic.tickClientData(Minecraft.getInstance());
     }
 
     @SubscribeEvent
